Move fields management to a new module (#48).
